+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 * Added a new option for the shortcode to define the id of the HTML container (useful when we have multiple shortcodes and we want to give a unique style to each instance).
 * Fixed applying CSS in "Page Builder" by SiteOrigin plugin.
+### Fixed
+* Fixed applying correct id to ul element when using the shortcode.
 
 ## [4.8.0] - 2019-06-01
 ### Added